feat: paginationのtotal_countの型を修正
- audit_grpc.rs, auth_grpc.rs, config_grpc.rs, tonic_service.rs: paginationのtotal_countをi64から元の型に修正
- config_handler.rs, config_schema_handler.rs: format!マクロの使用を修正

+++ regions/system/server/rust/auth/src/adapter/grpc/auth_grpc.rs
@@ -113,7 +113,7 @@ impl AuthGrpcService {
                 Ok(ListUsersResponse {
                     users: proto_users,
                     pagination: Some(PaginationResult {
-                        total_count: result.pagination.total_count as i64,
+                        total_count: result.pagination.total_count,
                         page: result.pagination.page,
                         page_size: result.pagination.page_size,
                         has_next: result.pagination.has_next,
